% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204142-R.R
\name{HMC}
\alias{HMC}
\title{Generate samples by Hamiltonian Monte Carlo}
\usage{
HMC(
  N,
  q_init,
  delta = 0.1,
  L = 10,
  logTarget,
  glogTarget,
  isRandom = FALSE,
  interval = c(0, 1)
)
}
\arguments{
\item{N}{number of samples}

\item{q_init}{vector of initial values}

\item{delta}{step size for \code{leapfrog}}

\item{L}{number of \code{leapfrog} steps in each iteration}

\item{logTarget}{function to calculate the log density of target distribution}

\item{glogTarget}{function to calculate the gradient of \code{logTarget}}

\item{isRandom}{logical value to determine whether to select parameter \code{delta} randomly}

\item{interval}{interval from which we select \code{delta} randomly}
}
\value{
a list concluding \code{q}, samples of target distribution of size \code{N} and \code{k}, the rejection count
}
\description{
This function runs the HMC algorithm provided the \code{logTarget} and gradient \code{glogTarget}.
}
\examples{
\dontrun{
  Sigma <- matrix(c(1,0.98,0.98,1), 2, 2)
  InvSigma <- solve(Sigma)
  logTarget <- function(x) {
    - t(x) \%*\% InvSigma \%*\% x / 2
  }
  glogTarget <- function(x) {
    - InvSigma \%*\% x
  }
  result <- HMC(10000, c(1, 0), 0.18, 20, logTarget, glogTarget)
  plot(result$q, xlab="x", ylab="y")
}
}
