% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MALA}
\alias{MALA}
\title{Generate samples by Metropolis Adjusted Langevin Algorithm}
\usage{
MALA(N, q_init, sigma, L, logTarget, glogTarget)
}
\arguments{
\item{N}{number of samples}

\item{q_init}{vector of initial values}

\item{sigma}{standard deviation of proposal distribution}

\item{L}{number of between-sample random numbers}

\item{logTarget}{function to calculate the log density of target distribution}

\item{glogTarget}{function to calculate the gradient of \code{logTarget}}
}
\value{
a list concluding \code{q}, samples of target distribution of size \code{N} and \code{k}, the rejection count
}
\description{
This function runs the Metropolis Adjusted Langevin Algorithm provided the \code{logTarget} and gradient \code{glogTarget}.
}
\examples{
\dontrun{
Sigma <- matrix(c(1,0.98,0.98,1), 2, 2)
InvSigma <- solve(Sigma)
logTarget <- function(x) {
  - t(x) \%*\% InvSigma \%*\% x / 2
}
glogTarget <- function(x) {
  - InvSigma \%*\% x
}
result <- MALA(10000, c(1, 0), 0.18, 20, logTarget, glogTarget)
plot(result$q, xlab="x", ylab="y")
}
}
