% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204142-R.R
\name{MH}
\alias{MH}
\title{Generate samples by Random-Walk Metropolis Algorithm}
\usage{
MH(
  N,
  q_init,
  sigma = 0.1,
  L = 1,
  logTarget,
  isRandom = FALSE,
  interval = c(0, 1)
)
}
\arguments{
\item{N}{number of samples}

\item{q_init}{vector of initial values}

\item{sigma}{standard deviation of proposal distribution}

\item{L}{number of between-sample random numbers}

\item{logTarget}{function to calculate the log density of target distribution}

\item{isRandom}{logical value to determine whether to select parameter \code{sigma} randomly}

\item{interval}{interval from which we select \code{sigma} randomly}
}
\value{
a list concluding \code{q}, samples of target distribution of size \code{N} and \code{k}, the rejection count
}
\description{
This function runs the Random Walk Metropolis Algorithm provided the \code{logTarget}.
}
\examples{
\dontrun{
  Sigma <- matrix(c(1,0.98,0.98,1), 2, 2)
  InvSigma <- solve(Sigma)
  logTarget <- function(x) {
    - t(x) \%*\% InvSigma \%*\% x / 2  
  }
  result <- MH(400, c(1, 0), 0.18, 20, logTarget)
  plot(result$q, xlab="x", ylab="y")
}
}
