% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204142-R.R
\name{NUTS}
\alias{NUTS}
\title{Generate samples by No U-Turn Sampler}
\usage{
NUTS(N, q_init, delta, logTarget, glogTarget)
}
\arguments{
\item{N}{number of samples}

\item{q_init}{vector of initial values}

\item{delta}{step size for \code{leapfrog}}

\item{logTarget}{function to calculate the log density of target distribution}

\item{glogTarget}{function to calculate the gradient of \code{logTarget}}
}
\value{
a matrix concluding samples of target distribution of size \code{N}
}
\description{
This function runs the NUTS algorithm provided the \code{logTarget} and gradient \code{glogTarget}.
}
\examples{
\dontrun{
  Sigma <- matrix(c(1,0.98,0.98,1), 2, 2)
  InvSigma <- solve(Sigma)
  logTarget <- function(x) {
    - t(x) \%*\% InvSigma \%*\% x / 2
  }
  glogTarget <- function(x) {
    - InvSigma \%*\% x
  }
  result <- NUTS(10000, c(1, 0), 0.18, logTarget, glogTarget)
  plot(result, xlab="x", ylab="y")
}
}
