\name{SA24204142-package}
\alias{SA24204142-package}
\docType{package}
\title{
  This package provides several function for MCMC.
}
\description{
This package provides four functions about MCMC. \code{MH} generates samples using Random-Walk Metropolis Algorithm. \code{MALA} generates samples using Metropolis Adjusted Langevin Algorithm. \code{HMC} generates samples using Hamiltonian Monte Carlo. \code{NUTS} generates samples using No U-Turn Sampler. These functions can be used to sample only from continuous distributions for which the density function can be evaluated (perhaps up to an unknown normalizing constant) and the partial derivatives of the log of the density function can be computed. 
}
\author{
SA24204142

Maintainer: SA24204142 <radian_cql@mail.ustc.edu.cn>
}

\keyword{ package }
